First quarto post on github
Add example code for the pmap function in R

This commit adds an example of how to use the pmap function from the purrr package in R. The code defines a simple function that adds three numbers together and creates a sample dataset consisting of a list of lists, where each inner list contains three numbers. The pmap function is used to apply the add_three function to each inner list in the dataset, resulting in a list of the sums. The resulting list is stored in the result variable and printed to the console. This example demonstrates the use of pmap to apply a function to multiple arguments in parallel and can be used as a starting point for similar tasks.